Move components out of modulear into processors
